arterys_sdk/requirements.txt
boto3==1.12.48
flask==2.0.3
numpy==1.21.0
opencv-python==4.3.0.36
Pillow==9.0.0
pydicom==1.4.2
pyyaml==5.4
requests==2.22.0
requests-toolbelt==0.9.1
SimpleITK==1.2.4
matplotlib==3.1.3